+++ changelog.md
@@ -5,4 +5,4 @@
 - preparing travis.yml
 - empty, stub modules
 - stack.yaml file
-
+- pre and postorder for trees